Move client/siteinfo packages to root folder.
